Agrega logica Back-end de CU04.

+++ DeSo/src/main/java/ddb/deso/almacenamiento/DAO/HabitacionDAO.java
@@ -18,5 +18,5 @@ public interface HabitacionDAO {
     void eliminarHabitacion(Habitacion habitacion);
     List<Habitacion> listar();
     List<Habitacion> listarPorTipo(TipoHab tipoHabitacion);
-    Habitacion buscarPorNumero(int numero);
+    Habitacion buscarPorNumero(Long numero);
 }
